requirements.txt
coreapi==2.3.3
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0